+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Deduplicate type names [#160](https://github.com/pulumi/pulumi-aws-native/issues/160)
 - [cf2pulumi] Fix supported check for config module [#165](https://github.com/pulumi/pulumi-aws-native/issues/165)
+- Improve inline schema type representation [#167](https://github.com/pulumi/pulumi-aws-native/pull/167)
 
 ## 0.1.0 (September 30, 2021)